Move the DC API variant into openid4vp, and stop duplicating the modes
Review point, and it exposed a worse version of itself: response mode
constants had ended up in two packages, and pkg/openid4vp - which owns
RequestObject.ResponseMode and the oneof that validates it - had none.

The four accepted modes now live beside that field and its validation, so
the constants and the accepted set cannot drift apart, and dcAPIVariant
becomes RequestObject.WithDCAPIResponseMode. Nothing about it was
verifier-specific: it took an openid4vp type, returned one, and set an
openid4vp field.

pkg/model keeps its two names as aliases rather than its own literals -
the resolver and its tests read better for having them, and openid4vp is
the lower layer (pkg/model already imports it; the reverse would be a
cycle).

The drift test moved with the function and still fails on injected drift
in its new home, which is the property worth keeping: the two channels
must be the same request seen through different delivery paths.
